+++ packages/design-system/src/components/Modal/Modal.tsx
@@ -1,7 +1,8 @@
 import React, { HTMLAttributes, ReactElement, ReactNode, useEffect, useRef } from 'react';
 import i18n from 'i18next';
 import { Dialog, DialogBackdrop, DialogDisclosure, useDialogState } from 'reakit/Dialog';
-import { IconName } from '@talend/icons';
+// eslint-disable-next-line @talend/import-depth
+import { IconName } from '@talend/icons/dist/typeUtils';
 
 import { ButtonDestructive, ButtonPrimary, ButtonSecondary } from '../Button';
 import { Icon } from '../Icon';