+++ packages/forms/src/deprecated/widgets/DatalistWidget/index.js
@@ -1,3 +1,0 @@
-import DatalistWidget from './DatalistWidget';
-
-export default DatalistWidget;