+++ packages/forms/src/deprecated/widgets/MultiSelectTagWidget/index.js
@@ -1,3 +1,0 @@
-import MultiSelectTagWidget from './MultiSelectTagWidget';
-
-export default MultiSelectTagWidget;